fix: role check implemented for realm and client

+++ src/handlers/client.rs
@@ -12,7 +12,10 @@ use crate::{
         token::TokenUser,
     },
     services::client::{delete_client_by_id, get_all_clients, get_client_by_id, insert_client, update_client_by_id},
-    utils::role_checker::{is_current_realm_admin, is_master_realm_admin},
+    utils::{
+        default_resource_checker::is_default_client,
+        role_checker::{is_current_realm_admin, is_master_realm_admin},
+    },
 };
 use axum::{extract::Path, Extension, Json};
 use sea_orm::prelude::Uuid;
@@ -20,6 +23,9 @@ use sea_orm::prelude::Uuid;
 pub async fn get_clients(user: TokenUser, Extension(state): Extension<Arc<AppState>>, Path(realm_id): Path<Uuid>) -> Result<Json<Vec<Model>>, Error> {
     if is_master_realm_admin(&user) || is_current_realm_admin(&user, &realm_id.to_string()) {
         let clients = get_all_clients(&state.db, realm_id).await?;
+        if clients.is_empty() {
+            return Err(Error::not_found());
+        }
         Ok(Json(clients))
     } else {
         Err(Error::Authenticate(AuthenticateError::NoResource))
@@ -75,6 +81,10 @@ pub async fn delete_client(
     Extension(state): Extension<Arc<AppState>>,
     Path((realm_id, client_id)): Path<(Uuid, Uuid)>,
 ) -> Result<Json<DeleteResponse>, Error> {
+    if is_default_client(client_id) {
+        return Err(Error::cannot_perform_operation("Cannot delete the default client"));
+    }
+
     if is_master_realm_admin(&user) || is_current_realm_admin(&user, &realm_id.to_string()) {
         let client = delete_client_by_id(&state.db, client_id).await?;
         Ok(Json(DeleteResponse {
